fixed assistant check in initialize
Was trying to avoid initialize for every call for URL and Document which was resonable.
However because we would probably want to send a new document in the same "session" we need to give the new document/url so that it can be added to the assistans store. This makes every url/document call build the assistant again and not reuse the old.